Rename `CommandAssigneeClass` to `AssigneeClass`
